Erstatter retry logikk for WebClient med Spring boot retryable + bump graphql version
