+++ pom.xml
@@ -35,8 +35,15 @@
         <token-support.version>2.1.9</token-support.version>
         <fpsoknad-felles.version>1.0.0</fpsoknad-felles.version>
 
+        <jacoco.version>0.8.8</jacoco.version>
         <pdfbox.version>2.0.27</pdfbox.version>
+        <nv-i18n.version>1.29</nv-i18n.version>
+        <javax.inject.version>1</javax.inject.version>
+        <reflections.version>0.10.2</reflections.version>
+        <logstash-logback-encoder.version>7.2</logstash-logback-encoder.version>
         <springdoc.version>1.6.14</springdoc.version>
+        <spring.vault.version>3.1.1</spring.vault.version>
+        <spring.graphql.version>2.0.1</spring.graphql.version>
     </properties>
     <scm>
         <connection>scm:git:https://github.com/navikt/fpsoknad-mottak.git</connection>
@@ -96,7 +103,7 @@
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
-                <version>0.8.8</version>
+                <version>${jacoco.version}</version>
                 <executions>
                     <execution>
                         <goals>
@@ -119,7 +126,7 @@
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
                 <artifactId>spring-cloud-vault-dependencies</artifactId>
-                <version>3.1.1</version>
+                <version>${spring.vault.version}</version>
                 <scope>import</scope>
                 <type>pom</type>
             </dependency>
@@ -181,7 +188,17 @@
         <dependency>
             <groupId>com.graphql-java-kickstart</groupId>
             <artifactId>graphql-webclient-spring-boot-starter</artifactId>
-            <version>2.0.1</version>
+            <version>${spring.graphql.version}</version>
+        </dependency>
+
+        <!-- Retry -->
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-aspects</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.retry</groupId>
+            <artifactId>spring-retry</artifactId>
         </dependency>
 
 
@@ -192,7 +209,7 @@
         <dependency>
             <groupId>javax.inject</groupId>
             <artifactId>javax.inject</artifactId>
-            <version>1</version>
+            <version>${javax.inject.version}</version>
         </dependency>
         <dependency>
             <groupId>org.aspectj</groupId>
@@ -205,13 +222,13 @@
         <dependency>
             <groupId>com.neovisionaries</groupId>
             <artifactId>nv-i18n</artifactId>
-            <version>1.29</version>
+            <version>${nv-i18n.version}</version>
         </dependency>
 
         <dependency>
             <groupId>net.logstash.logback</groupId>
             <artifactId>logstash-logback-encoder</artifactId>
-            <version>7.2</version>
+            <version>${logstash-logback-encoder.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.pdfbox</groupId>
@@ -267,7 +284,7 @@
         <dependency>
             <groupId>org.reflections</groupId>
             <artifactId>reflections</artifactId>
-            <version>0.10.2</version>
+            <version>${reflections.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
