+++ src/main/java/no/nav/foreldrepenger/mottak/MottakApplication.java
@@ -10,11 +10,13 @@ import org.springframework.kafka.annotation.EnableKafka;
 
 import no.nav.security.token.support.client.spring.oauth2.EnableOAuth2Client;
 import no.nav.security.token.support.spring.api.EnableJwtTokenValidation;
+import org.springframework.retry.annotation.EnableRetry;
 
 @EnableJwtTokenValidation(ignore = { "graphql.spring", "org.springframework", "org.springdoc" })
 @SpringBootApplication
 @EnableCaching
 @EnableKafka
+@EnableRetry
 @EnableOAuth2Client(cacheEnabled = true)
 @ConfigurationPropertiesScan
 public class MottakApplication {
