+++ src/main/java/no/nav/foreldrepenger/mottak/http/WebClientConfiguration.java
@@ -222,41 +222,4 @@ public class WebClientConfiguration {
         };
     }
 
-    @Component
-    public class TokenXExchangeFilterFunction implements ExchangeFilterFunction {
-
-        private static final Logger LOG = LoggerFactory.getLogger(TokenXExchangeFilterFunction.class);
-
-        private final OAuth2AccessTokenService service;
-        private final ClientConfigurationPropertiesMatcher matcher;
-        private final ClientConfigurationProperties configs;
-
-        TokenXExchangeFilterFunction(ClientConfigurationProperties configs, OAuth2AccessTokenService service, ClientConfigurationPropertiesMatcher matcher) {
-            this.service = service;
-            this.matcher = matcher;
-            this.configs = configs;
-        }
-
-        @Override
-        public Mono<ClientResponse> filter(ClientRequest req, ExchangeFunction next) {
-            var url = req.url();
-            var urlUtenQueryParam = url.toString().split("\\?")[0];
-            LOG.trace("Sjekker token exchange for {}", urlUtenQueryParam);
-            var config = matcher.findProperties(configs, url);
-            if (config.isPresent()) {
-                LOG.trace("Gjør token exchange for {} med konfig {}", urlUtenQueryParam, config);
-                var token = service.getAccessToken(config.get()).getAccessToken();
-                LOG.info("Token exchange for {} OK", urlUtenQueryParam);
-                return next.exchange(ClientRequest.from(req).header(AUTHORIZATION, BEARER + token)
-                    .build());
-            }
-            LOG.trace("Ingen token exchange for {}", urlUtenQueryParam);
-            return next.exchange(ClientRequest.from(req).build());
-        }
-
-        @Override
-        public String toString() {
-            return getClass().getSimpleName() + " [service=" + service + ", matcher=" + matcher + ", configs=" + configs + "]";
-        }
-    }
 }
