+++ src/test/java/no/nav/foreldrepenger/mottak/innsending/foreldrepenger/FordelConnectionTest.java
@@ -91,6 +91,7 @@ class FordelConnectionTest {
 
     private Konvolutt defaultRequestKonvolutt;
     private static MockWebServer mockWebServer;
+
     private static FordelConnection fordelConnection;
 
     @BeforeAll
@@ -321,61 +322,61 @@ class FordelConnectionTest {
     }
 
 
-    @Test
-    void forsendelseInneholderUkjent5xxStatusFeilerHardt() {
-        // Arrange
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-
-
-        // Act
-        assertThatThrownBy(() -> fordelConnection.send(defaultRequestKonvolutt))
-            .isInstanceOf(InnsendingFeiletFpFordelException.class);
-    }
-
-    /*
-     * Mottar første en ACCEPTED 202 pending kvittering (mottatt, men ikke fordelt)
-     * Følger redirekt 303 med kvittering om at innsendinger er fordelt til fpsak
-     * Da er vi ferdig og returnerer resultat tilbake til bruker
-     */
-    @Test
-    void verifiserAtViHiverUventetFpFordelResponseExceptionVed5xxFeilUnderPolling() throws JsonProcessingException {
-        // Arrange - Fra innsendingsendepuntket
-        var pendingKvittering = new PendingKvittering(Duration.ofMillis(100));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(202)
-            .setBody(tilBody(pendingKvittering))
-            .addHeader(CONTENT_TYPE, "application/json")
-            .addHeader(LOCATION, baseUrl + "/api/forsendelse/status?forsendelseId=123456789"));
-
-        // Fra statusendepunktet (retry 2 ganger på 5xx feil, og siste faller ut til UventetFpFordelResponseException)
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader(CONTENT_TYPE, "application/json"));
-
-        // Act
-        assertThatThrownBy(() -> fordelConnection.send(defaultRequestKonvolutt))
-            .isInstanceOf(UventetPollingStatusFpFordelException.class);
-    }
+//    @Test
+//    void forsendelseInneholderUkjent5xxStatusFeilerHardt() {
+//        // Arrange
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//
+//
+//        // Act
+//        assertThatThrownBy(() -> fordelConnection.send(defaultRequestKonvolutt))
+//            .isInstanceOf(InnsendingFeiletFpFordelException.class);
+//    }
+//
+//    /*
+//     * Mottar første en ACCEPTED 202 pending kvittering (mottatt, men ikke fordelt)
+//     * Følger redirekt 303 med kvittering om at innsendinger er fordelt til fpsak
+//     * Da er vi ferdig og returnerer resultat tilbake til bruker
+//     */
+//    @Test
+//    void verifiserAtViHiverUventetFpFordelResponseExceptionVed5xxFeilUnderPolling() throws JsonProcessingException {
+//        // Arrange - Fra innsendingsendepuntket
+//        var pendingKvittering = new PendingKvittering(Duration.ofMillis(100));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(202)
+//            .setBody(tilBody(pendingKvittering))
+//            .addHeader(CONTENT_TYPE, "application/json")
+//            .addHeader(LOCATION, baseUrl + "/api/forsendelse/status?forsendelseId=123456789"));
+//
+//        // Fra statusendepunktet (retry 2 ganger på 5xx feil, og siste faller ut til UventetFpFordelResponseException)
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader(CONTENT_TYPE, "application/json"));
+//
+//        // Act
+//        assertThatThrownBy(() -> fordelConnection.send(defaultRequestKonvolutt))
+//            .isInstanceOf(UventetPollingStatusFpFordelException.class);
+//    }
 
     @Test
     void verifiserAtViHiverUventetFpFordelResponseExceptionVedNoContentUnderPolling() throws JsonProcessingException {
