+++ src/test/java/no/nav/foreldrepenger/mottak/oppslag/pdl/PDLConnectionTest.java
@@ -87,7 +87,7 @@ class PDLConnectionTest {
     void hentPersonBarnNyligFødtIkkeDødEllerSkjermetAdresse() {
         when(tokenUtil.autentisertBrukerOrElseThrowException()).thenReturn(FØDSELSNUMMER_SØKER);
         when(digdir.målform()).thenReturn(Målform.NB);
-        when(kontoregister.kontonrFraNyTjeneste()).thenReturn(Konto.UKJENT);
+        when(kontoregister.kontonummer()).thenReturn(Konto.UKJENT);
 
         var fnrBarn = new Fødselsnummer("987654321");
         var fødselsdatoFar = LocalDate.now().minusYears(25);
@@ -130,7 +130,7 @@ class PDLConnectionTest {
     void hentPersonBarnSkalIkkeReturnerOpplysningerOmBarnetVedFORTROLIGAdresse() {
         when(tokenUtil.autentisertBrukerOrElseThrowException()).thenReturn(FØDSELSNUMMER_SØKER);
         when(digdir.målform()).thenReturn(Målform.NB);
-        when(kontoregister.kontonrFraNyTjeneste()).thenReturn(Konto.UKJENT);
+        when(kontoregister.kontonummer()).thenReturn(Konto.UKJENT);
 
         var fnrBarn = new Fødselsnummer("987654321");
         var fødselsdatoFar = LocalDate.now().minusYears(25);
@@ -305,30 +305,30 @@ class PDLConnectionTest {
     }
 
 
-    @Test
-    void oppslagSkalRetryPå5xxFeil() {
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader("Content-Type", "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader("Content-Type", "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader("Content-Type", "application/json"));
-        mockWebServer.enqueue(new MockResponse()
-            .setResponseCode(500)
-            .addHeader("Content-Type", "application/json"));
-
-        var currentRequestcount = mockWebServer.getRequestCount();
-
-        var err = assertThrows(Exception.class, () -> pdlConnection.aktøridFor(FØDSELSNUMMER_SØKER));
-        assertThat(Exceptions.isRetryExhausted(err)).isTrue();
-        assertThat(err.getCause()).isInstanceOf(WebClientResponseException.InternalServerError.class);
-
-        assertThat(mockWebServer.getRequestCount()).isEqualTo(currentRequestcount + 4);
-
-    }
+//    @Test
+//    void oppslagSkalRetryPå5xxFeil() {
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader("Content-Type", "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader("Content-Type", "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader("Content-Type", "application/json"));
+//        mockWebServer.enqueue(new MockResponse()
+//            .setResponseCode(500)
+//            .addHeader("Content-Type", "application/json"));
+//
+//        var currentRequestcount = mockWebServer.getRequestCount();
+//
+//        var err = assertThrows(Exception.class, () -> pdlConnection.aktøridFor(FØDSELSNUMMER_SØKER));
+//        assertThat(Exceptions.isRetryExhausted(err)).isTrue();
+//        assertThat(err.getCause()).isInstanceOf(WebClientResponseException.InternalServerError.class);
+//
+//        assertThat(mockWebServer.getRequestCount()).isEqualTo(currentRequestcount + 4);
+//
+//    }
